+++ .eslintrc.json
@@ -1,3 +1,10 @@
 {
-  "extends": ["next/core-web-vitals", "next/typescript"]
+    "extends": [
+        "next/core-web-vitals",
+        "next/typescript"
+    ],
+    "rules": {
+        "@typescript-eslint/array-type": "off",
+        "@typescript-eslint/no-explicit-any": "off"
+    }
 }
