+++ app/globals.css
@@ -1,38 +1,43 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
 
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
 @layer base {
-  a {
-    @apply text-blue-500 no-underline hover:underline;
-  }
-  html {
-    font-family: "Open Sans", system-ui, sans-serif;
-  }
+	a {
+		@apply text-blue-500 no-underline hover:underline;
+	}
+	html {
+		font-family: "Open Sans", system-ui, sans-serif;
+	}
+
+    main {
+        @apply p-30
+    }
 }
 
 @layer components {
-    .icon-button { @apply size-4 stroke-2} /* button that only contains an icon */
-    .icon-disabled {@apply stroke-slate-500}
-    .info-bubble { @apply text-blue-600 size-5} /* info bubble */
-    .info-link { 
-        @apply cursor-pointer underline decoration-dashed decoration-blue-600 underline-offset-4
-    }
-    /*.dropdown {
+	.icon-button {
+		@apply size-4 stroke-2;
+	} /* button that only contains an icon */
+	.icon-disabled {
+		@apply stroke-slate-500;
+	}
+	.info-bubble {
+		@apply text-blue-600 size-5;
+	} /* info bubble */
+	.info-link {
+		@apply cursor-pointer underline decoration-dashed decoration-blue-600 underline-offset-4;
+	}
+	/*.dropdown {
         
     }
 
     .dropdown-menu {
     }*/
-    .dropdown:focus-within .dropdown-menu {
-        @apply block; 
-        /* display:block; */
-    }
-    
+	.dropdown:focus-within .dropdown-menu {
+		@apply block;
+		/* display:block; */
+	}
 }
-
-
-
