+++ app/layout.tsx.orig
@@ -1,9 +1,18 @@
 import type { Metadata } from "next";
+import localFont from "next/font/local";
 import "./globals.css";
 
-// TODO: update tailwind imported fonts w/localFonts
-// see layout.tsx.orig
 
+const geistSans = localFont({
+  src: "./fonts/GeistVF.woff",
+  variable: "--font-geist-sans",
+  weight: "100 900",
+});
+const geistMono = localFont({
+  src: "./fonts/GeistMonoVF.woff",
+  variable: "--font-geist-mono",
+  weight: "100 900",
+});
 
 export const metadata: Metadata = {
   title: "NIAGADS Open Access API",
@@ -18,7 +27,8 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`antialiased`}>
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+      >
         {children}
       </body>
     </html>
