+++ app/view/table/[queryId]/TableWrapper.tsx
@@ -1,9 +1,9 @@
 'use client'
-import Table, {TableProps}  from "@/components/Table/Table";
+import Table, { TableProps } from "@/components/Table/Table";
 import { Alert } from "@/components/UI/Alert";
 import { Button } from "@/components/UI/Button";
 
-interface WrapperProps { 
+interface WrapperProps {
     endpoint: string,
     parameters: any
 }
@@ -18,12 +18,7 @@ export default function TableWrapper({ id, data, columns, options, endpoint, par
 
     return (
         <main>
-            <Alert variant="info" message="Server-side pagination not yet implemented.">
-                <div>
-                    <p>Displaying page XX out of XXX.</p>
-                    <p>To fetch paged data, increment the value of the <span className="font-medium text-red-600">page</span> parameter in your request.</p>
-                </div>
-            </Alert>
+
             <Button variant="secondary">Fetch Data</Button>
             <Table
                 id={id}
@@ -32,5 +27,5 @@ export default function TableWrapper({ id, data, columns, options, endpoint, par
                 options={options}
             />
         </main>
-    ) 
+    )
 }
